Showcase for some performance/memory optimizations.
Runs in 2083 milliseconds on my puny laptop.
Edge count is correct, but 5000 vertices are missing compared to the
description. Is their vertex count correct?
Graph loading is not really optimized (memory tradeoff and I don't have
enough time to do better). Removed idle wait time, because its
description in the stats was misleading.